requirements.txt
boto3
rasterio
geopandas
pillow
matplotlib
shapely
pyproj